Added features to Demo classifier

+++ Random/src/main/scala/Random.scala
@@ -171,7 +171,7 @@ object Random {
     val maximo = indexed_data.agg(max("featureIndex")).collect()(0)(0).toString.toDouble.toInt
     
     // Agrupamos y sumamos los counts por cada feature
-    val grouped_indexed_data = indexed_data.groupBy("device_id","label","featureIndex").agg(sum("count").as("count"))
+    val grouped_indexed_data = indexed_data.groupBy("device_id","label","featureIndex").agg(sum("count").cast("int").as("count"))
     // Agrupamos nuevamente y nos quedamos con la lista de features para cada device_id
     val grouped_data = grouped_indexed_data.groupBy("device_id","label").agg(collect_list("featureIndex").as("features"),collect_list("count").as("counts"))
